rename tagsRenderer file to tags-renderer
