+++ admin-dev/themes/new-theme/js/pages/product/components/categories/tags-renderer.ts
